Reorganized frontend files.
