+++ client/components/formComponents/tripReducer.js
