+++ client/components/headerPages/Register.jsx
@@ -11,6 +11,7 @@
 // Package dependencies
 import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
+
 // Components
 import Header from './Header';
