+++ client/components/itineraryComponents/itineraryReducer.js
