+++ client/components/itineraryComponents/suggestionsReducer.js
